SOLR-16203: Initialize schema plugins loaded by SPI name (#2175)
(cherry picked from commit 8201bf81d79b638f281c3d075ba7c5e240ee9e08)
